streamguard-ai/backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
alembic>=1.13,<2
pydantic>=2,<3
pydantic-settings>=2,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=4.0.1,<5
python-multipart>=0.0.9
redis>=5,<6
httpx>=0.27,<1
email-validator>=2,<3
aiokafka>=0.10.0,<1
websockets>=12.0,<13
